Compatible model to integrate calibration in EEPROM in order to extend a new model calibration based on Neurona
