+++ examples/TCS3200TickerCallback/TCS3200TickerCallback.ino
@@ -12,13 +12,14 @@ TCS3200 CS(TCS3200_S2, TCS3200_S3, TCS3200_OUT, TCS3200_S0, TCS3200_S1,
 
 void colorChanged(int color) {
 	if (CS.onChangeColor()) {
-		if (CS.onChangeCallback) CS.onChangeCallback(CS.readLastColorID());
+		if (CS.onChangeCallback)
+			CS.onChangeCallback(CS.readLastColorID());
 	}
 }
 
 void colorCallback(int color) {
-  Serial.print("Color cambiado a: ");
-  Serial.println(CS._ct[color].name);
+	Serial.print("Color cambiado a: ");
+	Serial.println(CS._ct[color].name);
 }
 
 Ticker CSTicker(colorChanged, CS.refreshTime, 0, MILLIS);
@@ -31,9 +32,12 @@ void setup() {
 	CS.setFrequency(0);
 	CS.nSamples(40);
 	CS.setRefreshTime(2000);
-	CS.loadBW();
-	CS.loadCT();
-  CS.setOnChangeCallback( colorCallback );
+	CS.voidBW(); // Load default Factory Black & White calibration
+	CS.voidCT(); // Load default Factory Color Table calibration
+
+	// CS.loadBW(); // Load EEPROM Black & White calibration
+	// CS.loadCT(); // Load EEPROM Color Table calibration
+	CS.setOnChangeCallback(colorCallback);
 }
 
 void loop() { CS.update(); }
